worker/requirements.txt
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
korail2==0.4.0
SRTrain==2.6.7
requests==2.31.0
fastapi==0.111.0
uvicorn==0.29.0